requirements.txt
Django==1.6.6
South==1.0
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.18
celery==3.1.13
certifi==14.05.14
dj-database-url==0.3.0
dj-static==0.0.6
django-celery==3.1.10
django-suit==0.2.9
elasticsearch==1.2.0
kombu==3.0.21
psycopg2==2.5.4
python-dateutil==2.2
python-decouple==2.3
python-ptrace==0.7
pytz==2014.7
redis==2.10.3
requests==2.4.0
six==1.7.3
static3==0.5.1
urllib3==1.9
waitress==0.8.9
wsgiref==0.1.2
gevent==1.0.1
greenlet==0.4.3
sendgrid==0.5.1
sendgrid-django==1.0.3
smtpapi==0.1.2
django-haystack==2.2.0
BeautifulSoup==3.2.1
feedparser==5.1.3
